undoing migrations

+++ server/models/expense.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Expense = sequelize.define('Expense', {
     userId: DataTypes.INTEGER,
     category: DataTypes.STRING,
-    amount: DataTypes.INTEGER,
+    amount: DataTypes.DECIMAL(10,2),
     description: DataTypes.STRING,
     date: DataTypes.STRING
   }, {});
